Move to ES6 modules in some pages.
