+++ js/README.md
@@ -1,0 +1,3 @@
+# js
+
+`js/*.js` are [JavaScript modules](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Guide/Modules).
