+++ open-attach.html
@@ -11,7 +11,8 @@
   <ul id="attachment-list">
     <li>Loading...</li>
   </ul>
-  <script src="scripts/open-attach.js"></script>
+
+  <script src="js/open-attach.js" type="module"> </script>
 </body>
 
 </html>
