+++ scripts/README.md
@@ -1,0 +1,3 @@
+# scripts
+
+`scripts/*.js` are traditional JavaScript files.
