--- scripts/declare-pimple.js
@@ -1,6 +1,0 @@
-// Declare `pimple` container powered by pimple.js
-
-(function () {
-  const self = this
-  this.pimple = new Pimple()
-}).apply(this)
